Rework API per previous PR comments
